arreglos

+++ PROJECTE UF3/Jordisaac/Jordisaac/UI.cpp
@@ -25,31 +25,31 @@ void UI::render()
 {
 	rHearts.x = MARGIN;
 	rHearts.y = MARGIN;
-	int nHearts = mPlayer->getStat(LIFE_CAPACITY) / 2;
-	float life = mPlayer->getHP();
+	int nHearts = (int)mPlayer->getStat(LIFE_CAPACITY) / 2;
+	float life = (float)mPlayer->getHP();
 	for (int i = 0; i < nHearts; i++)
 	{
 		if (life > 1)
 		{
-			iVideo->renderGraphic(gHeart, rHearts.x, rHearts.y, rHearts.w, rHearts.h, 0, 0, 1.5, 1.5);
+			iVideo->renderGraphic(gHeart, rHearts.x, rHearts.y, rHearts.w, rHearts.h, 0, 0, 1.5f, 1.5f);
 		}
 		else if (life == 1)
 		{
-			iVideo->renderGraphic(gHeart, rHearts.x, rHearts.y, rHearts.w, rHearts.h, rHearts.w, 0, 1.5, 1.5);
+			iVideo->renderGraphic(gHeart, rHearts.x, rHearts.y, rHearts.w, rHearts.h, rHearts.w, 0, 1.5f, 1.5f);
 		}
 		else
 		{
-			iVideo->renderGraphic(gHeart, rHearts.x, rHearts.y, rHearts.w, rHearts.h, rHearts.w * 1.5);
+			iVideo->renderGraphic(gHeart, rHearts.x, rHearts.y, rHearts.w, rHearts.h, (int)(rHearts.w * 1.5f));
 		}
 		life -= 2;
 		if (i == 5)
 		{
-			rHearts.y += rHearts.h * 1.5 - 10;
+			rHearts.y += (int)(rHearts.h * 1.5f - 10);
 			rHearts.x = MARGIN;
 		}
 		else
 		{
-			rHearts.x += rHearts.w * 1.5 - 10;
+			rHearts.x += (int)(rHearts.w * 1.5f - 10);
 		}
 	}
 }
